Translation of transcript
we can use this for translation of the transcript, but it is not the official google library. If this doesn't work can rather opt using the official api. 
Also we can use detect function for translation as well instead of defining such big dictionaries, just need to ensure the des language id proper
https://pypi.org/project/googletrans/
documentation link